feat: method get total user

+++ identity-service/src/main/java/authservice/service/IdentityService.java
@@ -37,5 +37,5 @@ public interface IdentityService {
 
   Boolean logout(HttpServletRequest request, HttpServletResponse response);
 
-  int getUserCount();
+  Long getUserCount();
 }
